bugfix: the average size of the documents should be "float" instead of "integer", I had to write it as "string" since there is no "writeFloat"

+++ trunk/lucene-AIP/src/java/org/apache/lucene/index/FilterIndexReader.java
@@ -202,7 +202,7 @@ public class FilterIndexReader extends IndexReader {
   
   //AIP change code (AVGL)
   @Override
-  public int avgDocSize() throws IOException{
+  public float avgDocSize() throws IOException{
       ensureOpen();
       return in.avgDocSize();
   }
